+++ src/components.d.ts
@@ -6,7 +6,7 @@
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { AlertTypes, LicenseTypeEnum, MenuItem as MenuItem1, TooltipPosition } from "./beans/index";
-import { AvatarSize, ButtonSizeEnum, ButtonVariantBean, CardVariants, ComboItemBean, DictionaryData, DividerOrientation, DividerSize, ExpandableListButtonAlign, ExpandableListStyle, HeaderUserData, InputStatusBean, InputTypeBean, ListDividerType, ListSize, MenuItem, NotificationType, OffCanvasVariantsEnum, PocketStatus, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, Size, SortDirection, StatusTagStatus, TabOrientationBean, TabSizeBean, ThemeVariant, ThemeVariantBean, ToastNotificationPositionsTypes, ToastNotificationTransitionTypes, ToastNotificationTypes, TransitionDirectionEnum, ZChipType, ZTableRowExpandedType, ZtoggleSwitchPositionEnum } from "./beans";
+import { AvatarSize, ButtonSizeEnum, ButtonVariantBean, ButtonVariantEnum, CardVariants, ComboItemBean, DictionaryData, DividerOrientation, DividerSize, ExpandableListButtonAlign, ExpandableListStyle, HeaderUserData, InputStatusBean, InputTypeBean, ListDividerType, ListSize, MenuItem, NotificationType, OffCanvasVariantsEnum, PocketStatus, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, Size, SortDirection, StatusTagStatus, TabOrientationBean, TabSizeBean, ThemeVariant, ThemeVariantBean, ToastNotificationPositionsTypes, ToastNotificationTransitionTypes, ToastNotificationTypes, TransitionDirectionEnum, ZChipType, ZFileUploadTypeEnum, ZSectionTitleDividerPosition, ZTableRowExpandedType, ZtoggleSwitchPositionEnum } from "./beans";
 import { ListItemBean } from "./beans/index.js";
 import { ZTypographyLevels } from "./components/typography/z-typography/index";
 export namespace Components {
@@ -322,6 +322,37 @@ export namespace Components {
           * [optional] Divider size
          */
         "size"?: DividerSize;
+    }
+    interface ZDragdropArea {
+    }
+    interface ZFile {
+        "fileNumber": number;
+    }
+    interface ZFileUpload {
+        /**
+          * Prop indicating the accepted file type: ex ".pdf, .doc, .jpg"
+         */
+        "acceptedFormat"?: string;
+        /**
+          * Prop indicating the button variant
+         */
+        "buttonVariant"?: ButtonVariantEnum;
+        /**
+          * Description
+         */
+        "description"?: string;
+        /**
+          * Max file dimension in Megabyte
+         */
+        "fileMaxSize"?: number;
+        /**
+          * Title
+         */
+        "mainTitle"?: string;
+        /**
+          * Prop indicating the file upload type - can be default or dragdrop
+         */
+        "type": ZFileUploadTypeEnum;
     }
     interface ZFooter {
         /**
@@ -1026,23 +1057,28 @@ export namespace Components {
     }
     interface ZNotification {
         /**
-          * action button text
+          * Action button text
          */
         "actiontext"?: string;
         /**
-          * icon on the left of the content
+          * Name of the icon on the left of the content
          */
         "contenticonname"?: string;
         /**
-          * enable close icon
+          * Enable close icon
          */
         "showclose"?: boolean;
         /**
-          * enable shadow
+          * Enable shadow.
+          * @deprecated shadow is available only for the `sticky` version of the notification.
          */
         "showshadow"?: boolean;
         /**
-          * alert variant type
+          * Enable sticky notification bar.
+         */
+        "sticky"?: boolean;
+        /**
+          * Alert variant type
          */
         "type": NotificationType;
     }
@@ -1216,6 +1252,16 @@ export namespace Components {
           * [optional] Show or hide arrow
          */
         "showArrow"?: boolean;
+    }
+    interface ZSectionTitle {
+        /**
+          * Divider position for the primary title. This prop only works if the secondary title is not set.
+         */
+        "dividerPosition": ZSectionTitleDividerPosition;
+        /**
+          * Whether the primary title text is uppercase.
+         */
+        "uppercase": boolean;
     }
     interface ZSelect {
         /**
@@ -1686,6 +1732,24 @@ declare global {
         prototype: HTMLZDividerElement;
         new (): HTMLZDividerElement;
     };
+    interface HTMLZDragdropAreaElement extends Components.ZDragdropArea, HTMLStencilElement {
+    }
+    var HTMLZDragdropAreaElement: {
+        prototype: HTMLZDragdropAreaElement;
+        new (): HTMLZDragdropAreaElement;
+    };
+    interface HTMLZFileElement extends Components.ZFile, HTMLStencilElement {
+    }
+    var HTMLZFileElement: {
+        prototype: HTMLZFileElement;
+        new (): HTMLZFileElement;
+    };
+    interface HTMLZFileUploadElement extends Components.ZFileUpload, HTMLStencilElement {
+    }
+    var HTMLZFileUploadElement: {
+        prototype: HTMLZFileUploadElement;
+        new (): HTMLZFileUploadElement;
+    };
     interface HTMLZFooterElement extends Components.ZFooter, HTMLStencilElement {
     }
     var HTMLZFooterElement: {
@@ -1986,6 +2050,12 @@ declare global {
         prototype: HTMLZPopoverElement;
         new (): HTMLZPopoverElement;
     };
+    interface HTMLZSectionTitleElement extends Components.ZSectionTitle, HTMLStencilElement {
+    }
+    var HTMLZSectionTitleElement: {
+        prototype: HTMLZSectionTitleElement;
+        new (): HTMLZSectionTitleElement;
+    };
     interface HTMLZSelectElement extends Components.ZSelect, HTMLStencilElement {
     }
     var HTMLZSelectElement: {
@@ -2149,6 +2219,9 @@ declare global {
         "z-contextual-menu": HTMLZContextualMenuElement;
         "z-cookiebar": HTMLZCookiebarElement;
         "z-divider": HTMLZDividerElement;
+        "z-dragdrop-area": HTMLZDragdropAreaElement;
+        "z-file": HTMLZFileElement;
+        "z-file-upload": HTMLZFileUploadElement;
         "z-footer": HTMLZFooterElement;
         "z-footer-link": HTMLZFooterLinkElement;
         "z-footer-section": HTMLZFooterSectionElement;
@@ -2199,6 +2272,7 @@ declare global {
         "z-pocket-header": HTMLZPocketHeaderElement;
         "z-pocket-message": HTMLZPocketMessageElement;
         "z-popover": HTMLZPopoverElement;
+        "z-section-title": HTMLZSectionTitleElement;
         "z-select": HTMLZSelectElement;
         "z-slideshow": HTMLZSlideshowElement;
         "z-status-tag": HTMLZStatusTagElement;
@@ -2571,6 +2645,49 @@ declare namespace LocalJSX {
           * [optional] Divider size
          */
         "size"?: DividerSize;
+    }
+    interface ZDragdropArea {
+        /**
+          * Emitted when user drop one or more files
+         */
+        "onFileDropped"?: (event: CustomEvent<any>) => void;
+    }
+    interface ZFile {
+        "fileNumber"?: number;
+        /**
+          * Emitted when a z-file component is removed from the DOM
+         */
+        "onRemoveFile"?: (event: CustomEvent<any>) => void;
+    }
+    interface ZFileUpload {
+        /**
+          * Prop indicating the accepted file type: ex ".pdf, .doc, .jpg"
+         */
+        "acceptedFormat"?: string;
+        /**
+          * Prop indicating the button variant
+         */
+        "buttonVariant"?: ButtonVariantEnum;
+        /**
+          * Description
+         */
+        "description"?: string;
+        /**
+          * Max file dimension in Megabyte
+         */
+        "fileMaxSize"?: number;
+        /**
+          * Title
+         */
+        "mainTitle"?: string;
+        /**
+          * Emitted when user select one or more files
+         */
+        "onFileInput"?: (event: CustomEvent<any>) => void;
+        /**
+          * Prop indicating the file upload type - can be default or dragdrop
+         */
+        "type"?: ZFileUploadTypeEnum;
     }
     interface ZFooter {
         /**
@@ -3372,31 +3489,36 @@ declare namespace LocalJSX {
     }
     interface ZNotification {
         /**
-          * action button text
+          * Action button text
          */
         "actiontext"?: string;
         /**
-          * icon on the left of the content
+          * Name of the icon on the left of the content
          */
         "contenticonname"?: string;
         /**
-          * notification action event
+          * Call to action clicked
          */
         "onNotificationAction"?: (event: CustomEvent<any>) => void;
         /**
-          * notification close event
+          * Close button clicked
          */
         "onNotificationClose"?: (event: CustomEvent<any>) => void;
         /**
-          * enable close icon
+          * Enable close icon
          */
         "showclose"?: boolean;
         /**
-          * enable shadow
+          * Enable shadow.
+          * @deprecated shadow is available only for the `sticky` version of the notification.
          */
         "showshadow"?: boolean;
         /**
-          * alert variant type
+          * Enable sticky notification bar.
+         */
+        "sticky"?: boolean;
+        /**
+          * Alert variant type
          */
         "type"?: NotificationType;
     }
@@ -3591,6 +3713,16 @@ declare namespace LocalJSX {
           * [optional] Show or hide arrow
          */
         "showArrow"?: boolean;
+    }
+    interface ZSectionTitle {
+        /**
+          * Divider position for the primary title. This prop only works if the secondary title is not set.
+         */
+        "dividerPosition"?: ZSectionTitleDividerPosition;
+        /**
+          * Whether the primary title text is uppercase.
+         */
+        "uppercase"?: boolean;
     }
     interface ZSelect {
         /**
@@ -3996,6 +4128,9 @@ declare namespace LocalJSX {
         "z-contextual-menu": ZContextualMenu;
         "z-cookiebar": ZCookiebar;
         "z-divider": ZDivider;
+        "z-dragdrop-area": ZDragdropArea;
+        "z-file": ZFile;
+        "z-file-upload": ZFileUpload;
         "z-footer": ZFooter;
         "z-footer-link": ZFooterLink;
         "z-footer-section": ZFooterSection;
@@ -4046,6 +4181,7 @@ declare namespace LocalJSX {
         "z-pocket-header": ZPocketHeader;
         "z-pocket-message": ZPocketMessage;
         "z-popover": ZPopover;
+        "z-section-title": ZSectionTitle;
         "z-select": ZSelect;
         "z-slideshow": ZSlideshow;
         "z-status-tag": ZStatusTag;
@@ -4094,6 +4230,9 @@ declare module "@stencil/core" {
             "z-contextual-menu": LocalJSX.ZContextualMenu & JSXBase.HTMLAttributes<HTMLZContextualMenuElement>;
             "z-cookiebar": LocalJSX.ZCookiebar & JSXBase.HTMLAttributes<HTMLZCookiebarElement>;
             "z-divider": LocalJSX.ZDivider & JSXBase.HTMLAttributes<HTMLZDividerElement>;
+            "z-dragdrop-area": LocalJSX.ZDragdropArea & JSXBase.HTMLAttributes<HTMLZDragdropAreaElement>;
+            "z-file": LocalJSX.ZFile & JSXBase.HTMLAttributes<HTMLZFileElement>;
+            "z-file-upload": LocalJSX.ZFileUpload & JSXBase.HTMLAttributes<HTMLZFileUploadElement>;
             "z-footer": LocalJSX.ZFooter & JSXBase.HTMLAttributes<HTMLZFooterElement>;
             "z-footer-link": LocalJSX.ZFooterLink & JSXBase.HTMLAttributes<HTMLZFooterLinkElement>;
             "z-footer-section": LocalJSX.ZFooterSection & JSXBase.HTMLAttributes<HTMLZFooterSectionElement>;
@@ -4144,6 +4283,7 @@ declare module "@stencil/core" {
             "z-pocket-header": LocalJSX.ZPocketHeader & JSXBase.HTMLAttributes<HTMLZPocketHeaderElement>;
             "z-pocket-message": LocalJSX.ZPocketMessage & JSXBase.HTMLAttributes<HTMLZPocketMessageElement>;
             "z-popover": LocalJSX.ZPopover & JSXBase.HTMLAttributes<HTMLZPopoverElement>;
+            "z-section-title": LocalJSX.ZSectionTitle & JSXBase.HTMLAttributes<HTMLZSectionTitleElement>;
             "z-select": LocalJSX.ZSelect & JSXBase.HTMLAttributes<HTMLZSelectElement>;
             "z-slideshow": LocalJSX.ZSlideshow & JSXBase.HTMLAttributes<HTMLZSlideshowElement>;
             "z-status-tag": LocalJSX.ZStatusTag & JSXBase.HTMLAttributes<HTMLZStatusTagElement>;